[fix/#59] enum 코드 리뷰
